+++ python/sglang/srt/server/engine.py
@@ -6,9 +6,11 @@ import logging
 import multiprocessing as mp
 import os
 import signal
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union, AsyncIterator
 
+import orjson
 import torch
+from fastapi import Request
 from sglang.srt.managers.data_parallel_controller import (
     run_data_parallel_controller_process,
 )
@@ -28,6 +30,7 @@ from sglang.srt.managers.tokenizer_manager import TokenizerManager
 from sglang.srt.openai_api.adapter import (
     load_chat_template_for_openai_api,
 )
+from sglang.srt.server.utils import create_error_response
 from sglang.srt.server_args import PortArgs, ServerArgs
 from sglang.srt.torch_memory_saver_adapter import TorchMemorySaverAdapter
 from sglang.srt.utils import (
@@ -41,6 +44,7 @@ from sglang.srt.utils import (
     set_ulimit,
 )
 from sglang.version import __version__
+from starlette.responses import StreamingResponse
 
 logger = logging.getLogger(__name__)
 
@@ -90,7 +94,7 @@ class Engine:
 
         # get the current event loop
         loop = asyncio.get_event_loop()
-        ret = loop.run_until_complete(generate_request(obj, None))
+        ret = loop.run_until_complete(self._generate_raw(obj, None))
 
         if stream is True:
 
@@ -139,7 +143,7 @@ class Engine:
             stream=stream,
         )
 
-        ret = await generate_request(obj, None)
+        ret = await self._generate_raw(obj, None)
 
         if stream is True:
             generator = ret.body_iterator
@@ -161,6 +165,35 @@ class Engine:
             return generator_wrapper()
         else:
             return ret
+
+    async def _generate_raw(self, obj: GenerateReqInput, request: Request):
+        if obj.stream:
+
+            async def stream_results() -> AsyncIterator[bytes]:
+                try:
+                    async for out in self.tokenizer_manager.generate_request(obj, request):
+                        yield b"data: " + orjson.dumps(
+                            out, option=orjson.OPT_NON_STR_KEYS
+                        ) + b"\n\n"
+                except ValueError as e:
+                    out = {"error": {"message": str(e)}}
+                    yield b"data: " + orjson.dumps(
+                        out, option=orjson.OPT_NON_STR_KEYS
+                    ) + b"\n\n"
+                yield b"data: [DONE]\n\n"
+
+            return StreamingResponse(
+                stream_results(),
+                media_type="text/event-stream",
+                background=self.tokenizer_manager.create_abort_task(obj),
+            )
+        else:
+            try:
+                ret = await self.tokenizer_manager.generate_request(obj, request).__anext__()
+                return ret
+            except ValueError as e:
+                logger.error(f"Error: {e}")
+                return create_error_response(e)
 
     def shutdown(self):
         kill_process_tree(os.getpid(), include_parent=False)
